+++ src/pages/components/image gallery/image-gallery.css
