+++ src/pages/hobbies/Cosplays.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import ImageGallery from "../components/ImageGallery";
+import ImageGallery from "../components/image gallery/ImageGallery";
 
 function Cosplays() {
   // image imports
