+++ src/pages/hobbies/Lego.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import ImageGallery from "../components/ImageGallery";
+import ImageGallery from "../components/image gallery/ImageGallery";
 
 function Lego() {
   // image imports
